Separate sample app and its sources
Move out the sources required to build some sample apps into a separate
directory so that only the necessary files are packaged in the `gonut`
binary.
