+++ assets/sample-apps-src/binary/clean.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+set -euo pipefail
+
+BASEDIR="$(cd "$(dirname "$0")" && pwd)"
+TARGET_DIR="${BASEDIR}/../../sample-apps/$(basename "${BASEDIR}")"
+
+if [[ -f "${TARGET_DIR}/binary" ]]; then
+  rm "${TARGET_DIR}/binary"
+fi
